html/cross-origin-opener-policy/reporting-coop-navigated-popup.https.html

<!doctype html>
<title>Cross-Origin-Opener-Policy: a navigated popup</title>
<!-- In particular this is different from coep-navigate-popup.https.html as this document initiates
     the navigation (and uses same-origin-allow-popups and no COEP as without that it cannot be
     observed). COOP should work identically, but implementations might have used the wrong
     authority. -->
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/utils.js"></script> <!-- Use token() to allow running tests in parallel -->
<script src="resources/reporting-common.js"></script>
<script>

promise_test( async t => {
  const popupName = token();
  const noCoopChannelName = token();
  const coopChannelName = token();
  reporting_test( (resolve) => {
    const noCOOP = `resources/coop-coep.py?coop=unsafe-none; report-to="${popupReportEndpoint.name}"&coep=&channel=${noCoopChannelName}`;
    const coop = `resources/coop-coep.py?coop=same-origin; report-to="${redirectReportEndpoint.name}&coep=&channel=${coopChannelName}`;

    const popup = window.open(noCOOP, popupName);
    const channel = new BroadcastChannel(coopChannelName);
    // Close the popup once the test is complete.
    // The browsing context is closed after the navigation hence use the broadcast channel
    // to trigger the closure.
    t.add_cleanup(() => {
      channel.postMessage("close");
    });
    popup.onload = t.step_func(() => {
      assert_equals(popup.name, popupName);
      assert_equals(new URL(popup.document.URL).pathname, noCOOP);
      channel.onmessage = t.step_func(event => {
        const payload = event.data;
        // The name should be empty, but we're checking the length rather than a
        // string comparison to "" to keep the random token out of error messages.
        assert_equals(payload.name.length, 0);
        assert_false(payload.opener);
        assert_true(popup.closed);
        resolve();
      });
      popup.location = coop;
    });
  },
  popupName,
  [
    // Reports expected for the navigation from "noCOOP" to "coop"
    {
      "endpoint": popupReportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "effective-policy": "unsafe-none",
          "navigation-uri": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
          "violation-type": "navigation-from-document"
        },
        "url": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
        "type": "coop"
      }
    },
    {
      "endpoint": redirectReportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
          "effective-policy": "same-origin",
          "navigation-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "violation-type": "navigation-to-document"
        },
        "url": RegExp(`coop-coep.py?.*channel=${coopChannelName}$`),
        "type": "coop"
      }
    },
  ],
  [
    // no reports expected for the popup's navigation from about:blank to "noCoop"
    {
      "endpoint": reportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": `${location.href}`,
          "effective-policy": "same-origin-allow-popups",
          "navigation-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "violation-type": "navigation-from-document"
        },
        "url": `${location.href}`,
        "type": "coop"
      }
    },
    {
      "endpoint": popupReportEndpoint,
      "report": {
        "body": {
          "disposition": "enforce",
          "document-uri": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
          "effective-policy": "unsafe-none",
          "navigation-uri": `${location.href}`,
          "violation-type": "navigation-to-document"
        },
        "url": RegExp(`coop-coep.py?.*channel=${noCoopChannelName}$`),
        "type": "coop"
      }
    },
  ]);
}, "Open a popup to a document without COOP, then navigate it to a document with");

</script>
